pyproject.toml
[tool.poetry]
name = "doc-mgmt"
version = "0.1.0"
description = ""
authors = ["Alexander Yanchenko <alexander.yanchenko@coverwhale.com>", "Bogdan Vlasyuk <bogdan.vlasyukcoverwhale.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.95.1"
uvicorn = "^0.22.0"
gunicorn = "^20.1.0"
pydantic = {extras = ["dotenv", "email"], version = "^1.10.7"}
dependency-injector = {extras = ["pydantic"], version = "^4.40.0"}
docxtpl = "^0.16.7"
httpx = "^0.24.1"
aioboto3 = "^11.2.0"
types-aioboto3-lite = {extras = ["essential"], version = "^11.2.0"}
aiofiles = "^23.1.0"
types-aiofiles = "^23.1.0"
pyhumps = "^3.7.3"
newrelic = "^8.7.1"
Jinja2 = "^3.1.2"
tenacity = "^8.2.2"
aiocache = "0.12.1"
docusign-esign = "3.22.0"
python-dateutil = "^2.8.2"
pypdf = {extras = ["image"], version = "3.7.1"}
python-multipart = "^0.0.6"
pypdftk = "^0.5"
pyjwt = {extras = ["crypto"], version = "^2.7.0"}


[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
pytest-cov = "^4.0.0"
bandit = "^1.7.5"
flake8-print = "^5.0.0"
pre-commit = "^3.3.1"
wemake-python-styleguide = "^0.17.0"
pytest-asyncio = "^0.21.0"
pytest-mock = "^3.9.0"
httpx = "^0.24.1"
asgi-lifespan = "^2.0.0"
pydantic = {extras = ["mypy"], version = "^1.10.7"}
types-aiofiles = "^23.1.0"
types-python-dateutil = "^2.8.19.13"
mypy = "^1.3.0"
typing-extensions = "^4.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
